pyproject.toml
[project]
name = "tapkit"
version = "0.1.0"
description = "Python SDK for controlling iPhones via TapKit"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27.0",
    "oagi",
    "pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["tapkit"]
